CAV2: MPKIL3 management and phase detection mechanism changed.
For the MPKIL3 of interval i of each app: 1. Check if valid_deque is equal to window size. Removed spike value detection and deque_aux. 2. Add MPKIL3 to v_mpkil3 and v_mpkil3_prev. 3. Check if app is isolated in CLOS 3 and if it is the case, return it to CLOS 1 if HPKIL3 > 1. 4. If phase_duration of app is > 1 calculate ICOV and if it is > 0.5, new phase is comming. Reset counters and associated variables. If app is critical, adjust the window_size to prev. phase duration (min. size 2 and max. size 10).